Initial commit of configuration library project
The configuration library will contain a class for each type of XML
file used by the game. They will be connected to provide for one-step
refactoring and one-click balancing. The basic idea is to create a
collection, fill it from the XML, transform it, and then create a new
XML document from the collection.

For now, only the XML files necessary for block transformations are
present, and they are all stubs.